requirements.txt
modelscope==1.9.5
transformers>=4.44.0
streamlit==1.24.0
sentencepiece==0.1.99
transformers_stream_generator==0.0.4
datasets==2.13.0
peft==0.10.0
openai==1.42.0
tqdm==4.64.1
setuptools==69.5.1
vllm==0.5.3
nest-asyncio
accelerate
tf-keras
autoawq
lmdeploy
loguru
pandas
faiss-cpu 
sentence-transformers